+++ src/routes/+layout.svelte
@@ -7,6 +7,8 @@
 	import Footer from '$lib/components/layout/Footer.svelte';
 	import CookieBanner from '$lib/components/layout/CookieBanner.svelte';
 	import JsonLd from '$lib/components/seo/JsonLd.svelte';
+
+
 </script>
 
 <!-- SITEWIDE JSON-LD -->